+++ HanoiTower/HanoiTowerWpf201/AppModel.cs
@@ -44,4 +44,13 @@ namespace HanoiTowerWpf201
 			Count.Value++;
 		}
 	}
+
+	public class Disk(int id, int towerId, int index)
+	{
+		public int Id { get; } = id;
+		public ReactiveProperty<int> TowerId { get; } = new ReactiveProperty<int>(towerId);
+		public ReactiveProperty<int> IndexInTower { get; } = new ReactiveProperty<int>(index);
+
+		public int Width => Id * 30 + 10;
+	}
 }
